fix: deprecate addField usage (category: user)
